+++ sdk/textanalytics/Azure.AI.TextAnalytics/CHANGELOG.md
@@ -6,6 +6,8 @@
 - In `LinkedEntity`, property `Id` has been renamed to `DataSourceEntityId`.
 - Added `DetectLanguageInput.None` for user convenience, instead of passing empty string to `CountryHint`.
 - Change wording in all documentation from `text inputs` to `documents`.
+- Properties `Length` and `Offset` have been renamed to `GraphemeLength` and `GraphemeOffset` for the `SentenceSentiment`,
+`CategorizedEntity`, `PiiEntity`, and `LinkedEntityMatch` objects.
 
 ## 1.0.0-preview.2 (2020-02-11)
 ### Breaking changes